+++ core/config.py
@@ -70,6 +70,7 @@ class ConfigManager:
         "disabled_current_thread_footer": "Please try again later...",
         "transfer_reactions": True,
         "close_on_leave": False,
+        "alert_on_mention": False,
         # moderation
         "recipient_color": str(discord.Color.gold()),
         "mod_color": str(discord.Color.green()),
@@ -144,6 +145,7 @@ class ConfigManager:
         "thread_move_notify_mods",
         "transfer_reactions",
         "close_on_leave",
+        "alert_on_mention",
         "enable_plugins",
         "enable_eval",
     }